refacto: locate thinc layers next to their component & fix docs
