+++ edsnlp/pipelines/trainable/nested_ner/__init__.py
@@ -1,0 +1,1 @@
+from .factory import create_component, create_scorer
