+++ edsnlp/pipelines/trainable/nested_ner/nested_ner.py
@@ -10,7 +10,6 @@ from spacy.training import Example
 from spacy.vocab import Vocab
 from thinc.api import Model, Optimizer
 from thinc.backends import NumpyOps
-from thinc.config import Config
 from thinc.model import set_dropout_rate
 from thinc.types import Ints2d
 from wasabi import Printer
@@ -19,68 +18,9 @@ from edsnlp.utils.filter import filter_spans
 
 msg = Printer()
 
-
 NUM_INITIALIZATION_EXAMPLES = 1000
 
-
-nested_ner_default_config = """
-[model]
-    @architectures = "eds.stack_crf_ner_model.v1"
-    mode = "joint"
-
-    [model.tok2vec]
-        @architectures = "spacy.Tok2Vec.v1"
-
-    [model.tok2vec.embed]
-        @architectures = "spacy.MultiHashEmbed.v1"
-        width = 96
-        rows = [5000, 2000, 1000, 1000]
-        attrs = ["ORTH", "PREFIX", "SUFFIX", "SHAPE"]
-        include_static_vectors = false
-
-    [model.tok2vec.encode]
-        @architectures = "spacy.MaxoutWindowEncoder.v1"
-        width = ${model.tok2vec.embed.width}
-        window_size = 1
-        maxout_pieces = 3
-        depth = 4
-
-[scorer]
-    @scorers = "eds.nested_ner_scorer.v1"
-"""
-
-NESTED_NER_DEFAULTS = Config().from_str(nested_ner_default_config)
 np_ops = NumpyOps()
-
-
-@Language.factory(
-    "nested_ner",
-    default_config=NESTED_NER_DEFAULTS,
-    requires=["doc.ents", "doc.spans"],
-    assigns=["doc.ents", "doc.spans"],
-    default_score_weights={
-        "ents_f": 1.0,
-        "ents_p": 0.0,
-        "ents_r": 0.0,
-    },
-)
-def create_component(
-    nlp: Language,
-    name: str,
-    model: Model,
-    ent_labels=None,
-    spans_labels=None,
-    scorer=None,
-):
-    """Construct a TrainableQualifier component."""
-    return TrainableNer(
-        vocab=nlp.vocab,
-        model=model,
-        name=name,
-        ent_labels=ent_labels,
-        spans_labels=spans_labels,
-        scorer=scorer,
-    )
 
 
 def nested_ner_scorer(examples: Iterable[Example], **cfg):
